+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+## [0.5.12] 2024-02-24
+- Happy Birthday Ruby!
+- Added --next CLI option
+- Added --pipeline CLI option
+
 ## [0.5.11] 2024-02-18
 - allow directives to return information that is inserted into the prompt text
 - added //shell command directive